requirements.txt
django-cors-headers==4.3.1
django-jazzmin==2.6.0
gunicorn==21.2.0
pip-chill==1.0.3
psycopg2-binary==2.9.9
python-decouple==3.8
